Move file to wrong place
