+++ src/majors.ts
